+++ CHANGES.rst
@@ -1,7 +1,15 @@
 0.3 (unreleased)
 ----------------
 
-- No changes yet.
+- Added support for the ``c=`` argument, similar to the argument with
+  the same name in Matplotlib's ``scatter`` function.
+
+- Added support for log axes.
+
+- Fixed support for flipped limits (e.g. xmax < xmin)
+
+- Added support for setting vmin/vmax to functions to determine limits
+  on-the-fly.
 
 0.2 (2017-07-20)
 ----------------